Split question DTOs into request and response subfolders
